vector_db/requirements.txt
accelerate==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bracex==2.6
build==1.4.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.2
colorlog==6.10.1
cryptography==46.0.7
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
deepagents==0.5.3
defusedxml==0.7.1
dill==0.4.1
distro==1.9.0
docling==2.90.0
docling-core==2.74.0
docling-ibm-models==3.13.0
docling-parse==5.9.0
docstring-parser==0.18.0
durationpy==0.10
et-xmlfile==2.0.0
faker==40.15.0
filelock==3.28.0
filetype==1.2.0
flatbuffers==25.12.19
fsspec==2026.3.0
google-auth==2.49.2
google-genai==1.73.1
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.11.0
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.1
importlib-resources==7.1.0
jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.15
langchain-anthropic==1.4.1
langchain-core==1.3.0
langchain-google-genai==4.2.2
langchain-ollama==1.1.0
langgraph==1.1.8
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.33
latex2mathml==3.81.0
lxml==6.1.0
markdown-it-py==4.0.0
marko==2.2.2
markupsafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.19
networkx==3.6.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
oauthlib==3.3.1
ollama==0.6.1
omegaconf==2.3.0
onnxruntime==1.24.4
openai==2.32.0
opencv-python==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.1
pandas==3.0.2
pdf2image==1.17.0
pdfminer-six==20251230
pdfplumber==0.11.9
pillow==12.2.0
pluggy==1.6.0
polyfactory==3.3.0
portalocker==3.2.0
protobuf==6.33.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1-modules==0.4.2
pybase64==1.4.3
pyclipper==1.4.0
pycparser==3.0
pydantic==2.13.2
pydantic-core==2.46.2
pydantic-settings==2.13.1
pygments==2.20.0
pylatexenc==2.10
pypdfium2==5.7.0
pypika==0.51.1
pyproject-hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-pptx==1.0.2
pyyaml==6.0.3
qdrant-client==1.17.1
rapidocr==3.8.1
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
rtree==1.4.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semchunk==3.2.5
sentence-transformers==5.4.1
setuptools==81.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.4
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
triton==3.6.0
typer==0.21.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
wcmatch==10.1
websocket-client==1.9.0
websockets==16.0
xlsxwriter==3.2.9
xxhash==3.6.0
zipp==3.23.1
zstandard==0.25.0
